Simplify heuristics and reduce indentation
Reduced indentation by returning early when there are no results.
Combined heuristics 1-5 into a single heuristic that uses frequency
to score based on the number of layers.
Technique names are added as signatures to allow per technique scoring
in the future but for now they all score 10 to be in line with the original
scoring.